+++ src/onegov.event/HISTORY.rst
@@ -13,6 +13,9 @@ Unreleased
 - Add autoclean option to add event function.
   [msom]
 
+- Add by_id method for event collections.
+  [msom]
+
 0.0.1 (2015-08-20)
 ~~~~~~~~~~~~~~~~~~~